requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
customtkinter==5.2.1
cycler==0.12.1
Cython==3.0.6
darkdetect==0.8.0
ffmpeg==1.4
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.45.1
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
idna==3.6
Jinja2==3.1.2
keras==2.14.0
kiwisolver==1.4.5
lapx==0.5.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.25.1
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thop==0.1.1.post2209072238
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.219
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
